--- Hello world
@@ -1,5 +1,0 @@
-#include<stdio.h>
-int main()
-{
-  printf("Hello world\n");
-}
